Fix warning: implicit conversion changes signedness
visp/tutorial/image/tutorial-draw-text.cpp: In function ‘int main()’:
visp/tutorial/image/tutorial-draw-text.cpp:22:48: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
   22 |   vpDisplay::displayRectangle(I, I.getHeight() / 4, I.getWidth() / 4, I.getWidth() / 2, I.getHeight() / 2,
      |                                  ~~~~~~~~~~~~~~^~~
visp/tutorial/image/tutorial-draw-text.cpp:22:66: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
   22 |   vpDisplay::displayRectangle(I, I.getHeight() / 4, I.getWidth() / 4, I.getWidth() / 2, I.getHeight() / 2,
      |                                                     ~~~~~~~~~~~~~^~~
visp/tutorial/image/tutorial-draw-text.cpp:25:43: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
   25 |   vpDisplay::displayText(I, I.getHeight() / 2, I.getWidth() / 2, "Hello World!", vpColor::yellow);
      |                             ~~~~~~~~~~~~~~^~~
visp/tutorial/image/tutorial-draw-text.cpp:25:61: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
   25 |   vpDisplay::displayText(I, I.getHeight() / 2, I.getWidth() / 2, "Hello World!", vpColor::yellow);
      |                                                ~~~~~~~~~~~~~^~~

+++ tutorial/image/tutorial-draw-text.cpp
@@ -19,10 +19,10 @@ int main()
 #endif
 
   vpDisplay::display(I);
-  vpDisplay::displayRectangle(I, I.getHeight() / 4, I.getWidth() / 4, I.getWidth() / 2, I.getHeight() / 2,
-                              vpColor::red, true);
+  vpDisplay::displayRectangle(I, static_cast<int>(I.getHeight()) / 4, static_cast<int>(I.getWidth()) / 4,
+                              I.getWidth() / 2, I.getHeight() / 2, vpColor::red, true);
   //! [text]
-  vpDisplay::displayText(I, I.getHeight() / 2, I.getWidth() / 2, "Hello World!", vpColor::yellow);
+  vpDisplay::displayText(I, static_cast<int>(I.getHeight()) / 2, static_cast<int>(I.getWidth()) / 2, "Hello World!", vpColor::yellow);
   //! [text]
   vpDisplay::flush(I);
   std::cout << "A click to quit..." << std::endl;
